Only reconfigure what we need

+++ py/liquid/reconfigure.py
@@ -1,4 +1,5 @@
 import sys
+import os
 import json
 from pathlib import Path
 import subprocess
@@ -9,6 +10,7 @@ from .vpn import client
 from . import discover
 
 GLOBAL_LIQUID_OPTIONS = '/var/lib/liquid/conf/options.json'
+CURRENT_LIQUID_VARS = '/var/lib/liquid/conf/vars.json'
 
 
 def get_liquid_options():
@@ -16,15 +18,29 @@ def get_liquid_options():
         return json.load(f)
 
 
+def get_current_vars():
+    if not os.path.exists(CURRENT_LIQUID_VARS):
+        return {}
+    with open(CURRENT_LIQUID_VARS, encoding='utf8') as f:
+        return json.load(f)
+
+
+def write_current_vars(vars):
+    tmp = CURRENT_LIQUID_VARS + '.tmp'
+    with open(tmp, 'w', encoding='utf8') as f:
+        print(json.dumps(vars), file=f)
+    os.rename(tmp, CURRENT_LIQUID_VARS)
+
+
 def run(cmd):
     print('+', cmd)
     subprocess.run(cmd, shell=True, check=True)
 
 
-def ansible(vars):
+def ansible(vars, tags):
     builder = Builder_cloud()
     (builder.setup / 'ansible' / 'vars' / 'config.yml').touch()
-    builder.update('configure', None, vars)
+    builder.update(tags, None, vars)
 
 
 def on_reconfigure():
@@ -33,26 +49,70 @@ def on_reconfigure():
     vars = {'liquid_{}'.format(k): v for k, v in options['vars'].items()}
     vars['liquid_apps'] = get_liquid_options().get('apps', True)
 
-    ansible(vars)
-    run('/opt/common/initialize.sh')
+    old_vars = get_current_vars()
+    first_boot = not old_vars
 
-    print('configure_lan')
-    configure_lan(vars)
+    print('old_vars:', json.dumps(old_vars))
+    print('vars:', json.dumps(vars))
 
-    print('configure_wifi')
-    configure_wifi(vars)
+    changes = set()
 
-    print('syncing vpn client keys')
-    client.sync_keys(vars)
+    if vars['liquid_lan'] != old_vars.get('liquid_lan'):
+        changes.add('lan')
 
-    print('configuring avahi interfaces')
-    discover.configure_avahi(vars)
+    if vars['liquid_wan'] != old_vars.get('liquid_wan'):
+        changes.add('wan')
 
-    if vars['liquid_ssh']['enabled']:
-        run('systemctl reload-or-restart ssh')
+    if vars['liquid_ssh'] != old_vars.get('liquid_ssh'):
+        changes.add('ssh')
+
+    if vars['liquid_vpn'] != old_vars.get('liquid_vpn'):
+        changes.add('vpn')
+
+    if vars['liquid_services'] != old_vars.get('liquid_services'):
+        changes.add('apps')
+
+    print('changes:', changes)
+
+    if first_boot:
+        tags = 'configure'
+
     else:
-        run('systemctl stop ssh')
+        tags = ','.join('configure-{}'.format(c) for c in changes)
 
-    run('service nginx restart')
-    run('supervisorctl update')
-    run('supervisorctl restart all')
+    print('tags:', tags)
+
+    if tags:
+        ansible(vars, tags)
+
+    if changes.intersection({'apps'}):
+        run('/opt/common/initialize.sh')
+
+    if changes.intersection({'lan'}):
+        print('configure_lan')
+        configure_lan(vars)
+
+    if changes.intersection({'lan', 'wan'}):
+        print('configure_wifi')
+        configure_wifi(vars)
+
+    if changes.intersection({'vpn'}):
+        print('syncing vpn client keys')
+        client.sync_keys(vars)
+
+    if changes.intersection({'lan', 'wan', 'vpn'}):
+        print('configuring avahi interfaces')
+        discover.configure_avahi(vars)
+
+    if changes.intersection({'ssh'}):
+        if vars['liquid_ssh']['enabled']:
+            run('systemctl reload-or-restart ssh')
+        else:
+            run('systemctl stop ssh')
+
+    if changes.intersection({'apps'}):
+        run('service nginx restart')
+        run('supervisorctl update')
+        run('supervisorctl restart all')
+
+    write_current_vars(vars)
